backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
alembic==1.12.1
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
email-validator==2.2.0
